footer summary 변경 작업 (css변경 전)
